Factor out StoppableThread
